HackStudio: Reorganize debugger-related files
The debugger's files are now placed under HackStudio/Debugger
